+++ VAMobile/src/translations/en/common.json
@@ -863,24 +863,29 @@
   "howWillYou.useInfo.2": "If you’re enrolled in VA health care, we’ll send your prescriptions to your mailing address. Your health care team may also use this information to contact you.",
   "icon.success": "Success icon,",
   "inAppFeedback.a11yHint": "Go to the give feedback page",
-  "inAppFeedback.popup.title": "Give feedback on the app?",
-  "inAppFeedback.popup.body": "We'll use your feedback to improve the app. Your feedback stays confidential.",
-  "inAppFeedback.popup.notNow": "Not now",
-  "inAppFeedback.whatTask.header": "What task were you trying to do today?",
-  "inAppFeedback.whatTask.body": "Please don't enter personal information into the text box. This includes your name, address, Social Security Number, medical information or anything else that someone could use to identify you.",
-  "inAppFeedback.taskSelection.header": "What task were you trying to do today?",
-  "inAppFeedback.overallSatisfaction.header": "What is your overall satisfaction with this app?",
+  "inAppFeedback.agreeOrDisagree.header": "2. Please let us know how much you agree or disagree with the following statements:",
+  "inAppFeedback.agreeOrDisagree.stronglyAgree": "Strongly agree",
+  "inAppFeedback.agreeOrDisagree.agree": "Agree",
+  "inAppFeedback.agreeOrDisagree.neither": "Neither agree or disagree",
+  "inAppFeedback.agreeOrDisagree.disagree": "Disagree",
+  "inAppFeedback.agreeOrDisagree.stronglyDisagree": "Strongly Disagree",
+  "inAppFeedback.easyToUse.header": "b. The VA mobile app is easy to use.",
+  "inAppFeedback.functions.header": "a. VA mobile app's functions meet my needs.",
+  "inAppFeedback.legalReqs.paragraph": "VA may utilize individual Veteran survey data from this survey or other sources to ensure the final scores truly and accurately represent the experiences of Veterans. This information is collected in accordance with section 3507 of the Paperwork Reduction Act of 1995. Title 38, United States Code, allows us to ask for this information. We estimate that you will need an average of 3 minutes to review the instructions and complete this survey. The results of this survey will be used to inform opportunities for program improvement in the quality of VA services. Participation in this survey is voluntary, and your decision not to respond will have no impact on VA benefits or services which you may currently be receiving. VA cannot conduct or sponsor a collection of information unless a valid OMB control number is displayed. You are not required to respond to a collection of information if this number is not displayed.",
+  "inAppFeedback.legalReqs.paragraph.2": "Information gathered will be kept private to the extent provided by law.",
+  "inAppFeedback.legalReqs.paragraph.link": "Get valid control numbers on OMB website",
+  "inAppFeedback.overallSatisfaction.header": "1. What is your overall satisfaction with the VA mobile app?",
   "inAppFeedback.overallSatisfaction.notAtAllSatisfied": "Not at all satisfied",
   "inAppFeedback.overallSatisfaction.dissatisfied": "Dissatisfied",
   "inAppFeedback.overallSatisfaction.neither": "Neither satisfied or dissatisfied",
   "inAppFeedback.overallSatisfaction.satisfied": "Satisfied",
   "inAppFeedback.overallSatisfaction.verySatisfied": "Very satisfied",
-  "inAppFeedback.legalReqs.number": "OMB Number: 2900-0876",
-  "inAppFeedback.legalReqs.expiration": "Expiration: 02/28/2026",
-  "inAppFeedback.legalReqs.burdenTime": "Burden Time: 3 Minutes",
-  "inAppFeedback.legalReqs.paragraph": "VA may utilize individual Veteran survey data from this survey or other sources to ensure the final scores truly and accurately represent the experiences of Veterans. This information is collected in accordance with section 3507 of the Paperwork Reduction Act of 1995. Title 38, United States Code, allows us to ask for this information. We estimate that you will need an average of 3 minutes to review the instructions and complete this survey. The results of this survey will be used to inform opportunities for program improvement in the quality of VA services. Participation in this survey is voluntary, and your decision not to respond will have no impact on VA benefits or services which you may currently be receiving. VA cannot conduct or sponsor a collection of information unless a valid OMB control number is displayed. You are not required to respond to a collection of information if this number is not displayed.",
-  "inAppFeedback.legalReqs.paragraph.2": "Information gathered will be kept private to the extent provided by law.",
-  "inAppFeedback.legalReqs.paragraph.link": "Get valid control numbers on OMB website",
+  "inAppFeedback.popup.title": "Give feedback on the app?",
+  "inAppFeedback.popup.body": "We'll use your feedback to improve the app. Your feedback stays confidential.",
+  "inAppFeedback.popup.notNow": "Not now",
+  "inAppFeedback.taskSelection.header": "What task were you trying to do today?",
+  "inAppFeedback.whatTask.header": "3. What feedback, if any, do you have about your experience with the app? (Optional)",
+  "inAppFeedback.whatTask.body": "Please don't enter personal information such as name, Social Security number, medical information, or anything else someone could use to identify you.",
   "inAppFeedback.submitFeedback": "Submit feedback",
   "inAppFeedback.personalInfo.title": "Submit your feedback with personal information?",
   "inAppFeedback.personalInfo.body": "You can still submit your feedback. But we recommend deleting any personal information you included.",
